Refactor microfrontends into dedicated route modules
