+++ src/microfrontends/operations-reports/client/routes/context.ts
@@ -1,0 +1,12 @@
+import { useOutletContext } from 'react-router-dom';
+
+import type { OperationsReport } from '../types';
+
+export type ReportsRouteContext = {
+  reports: OperationsReport[];
+  isLoading: boolean;
+  error: string | null;
+};
+
+export const useReportsRouteContext = (): ReportsRouteContext =>
+  useOutletContext<ReportsRouteContext>();
