+++ src/microfrontends/users-and-roles/client/routes/context.ts
@@ -1,0 +1,11 @@
+import { useOutletContext } from 'react-router-dom';
+
+import type { UserRecord } from '../types';
+
+export type UsersRouteContext = {
+  users: UserRecord[];
+  isLoading: boolean;
+  error: string | null;
+};
+
+export const useUsersRouteContext = (): UsersRouteContext => useOutletContext<UsersRouteContext>();
