+++ src/microfrontends/users-and-roles/client/routes/utils.ts
@@ -1,0 +1,7 @@
+import type { UserStatus } from '../types';
+
+export const statusColors: Record<UserStatus, string> = {
+  Active: 'green',
+  Invited: 'geekblue',
+  Suspended: 'volcano',
+};
